requirements.txt
altair==5.1.2
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.1.3
python-dotenv==1.0.0
selenium==4.14.0
sqlalchemy==2.0.21
streamlit==1.28.0
tqdm==4.66.1
vaderSentiment==3.3.2

